Fixed always failing validation in control panel on Plone 6.
The fields in the schemas for datagrid rows must not be required.
Otherwise the fields in the always present empty row at the bottom are required as well, and always give a validation error.
In Plone 5.2 this was fine, I think this may have been a change in zope.schema.